refactor(usage): keep model price management on web and desktop
