+++ apps/web/src/components/usage/usagePriceForm.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vite-plus/test";
 
-import { parseUsagePriceForm, usagePriceForm } from "./usagePriceOverrides.ts";
+import { parseUsagePriceForm, usagePriceForm } from "./usagePriceForm.ts";
 
 describe("model price entry", () => {
   it("preserves exact model IDs and fractional rates", () => {
