+++ packages/shared/package.json
@@ -243,10 +243,6 @@
       "types": "./src/usageFormat.ts",
       "import": "./src/usageFormat.ts"
     },
-    "./usagePriceOverrides": {
-      "types": "./src/usagePriceOverrides.ts",
-      "import": "./src/usagePriceOverrides.ts"
-    },
     "./usageLimits": {
       "types": "./src/usageLimits.ts",
       "import": "./src/usageLimits.ts"
